Remove document hidden check
We should let the user decided if he want to show the notification when the document visible too.
Specially because document.hidden is only true for when minimized or tab is switched. If the user just alt-tab applications, the notification will never be displayed.

I suppose a common solution would be to check for window.document.hasFocus() instead, but, then again, I think we should leave this to the user.